added some error handling for files that already exist

+++ bin/svg2png-cli.js
@@ -35,4 +35,8 @@ const input = fs.readFileSync(argv._[0]);
 const output = svg2png.sync(input, { width: argv.width, height: argv.height });
 
 const outputFilename = argv.output || path.basename(argv._[0], ".svg") + ".png";
-fs.writeFileSync(outputFilename, output, { flag: "wx" });
+
+if (!fs.existsSync(outputFilename)) {
+    fs.writeFileSync(outputFilename, output, { flag: "wx" });
+}
+
